Refactored datastore type specific files to be generic fetchers
