+++ src/datastore/contexts/CoreStoryProvider.tsx
@@ -3,8 +3,8 @@ import { useRegistry, useRepositoryEvents } from '@/hooks'
 import { useSession } from 'next-auth/react'
 import { useIndexedDB } from '../hooks/useIndexedDB'
 import { fetchOrRefresh } from '../lib/fetchOrRefresh'
-import { fetchStories } from '../lib/fetchStories'
 import { type IDBStory } from '../types'
+import { type IndexedStory } from '@/lib/index'
 
 interface CoreStoryProviderState {
   objects: IDBStory[]
@@ -31,12 +31,45 @@ export const CoreStoryProvider = ({ children }: {
       return
     }
 
-    const cachedObjects = await fetchOrRefresh<IDBStory>(
+    const cachedObjects = await fetchOrRefresh<IDBStory, IndexedStory>(
       IDB,
       documentType,
+      indexUrl,
+      data.accessToken,
       force,
-      async () => {
-        return await fetchStories(indexUrl, data.accessToken)
+      (item) => {
+        const { _id: id, _source: _ } = item
+        const getRoleText = (roleIndex: number): [string, string] => ([
+          _['document.meta.core_definition.role']?.[roleIndex]?.trim() || '',
+          _['document.meta.core_definition.data.text']?.[roleIndex]?.trim() || ''
+        ])
+        const [role0, text0] = getRoleText(0)
+        const [role1, text1] = getRoleText(1)
+
+        const story = {
+          id,
+          title: _['document.title'][0].trim(),
+          shortText: '',
+          longText: ''
+        }
+
+        if (role0 && text0) {
+          if (role0 === 'short') {
+            story.shortText = text0
+          } else if (role0 === 'long') {
+            story.longText = text0
+          }
+        }
+
+        if (role1 && text1) {
+          if (role1 === 'short') {
+            story.shortText = text1
+          } else if (role1 === 'long') {
+            story.longText = text1
+          }
+        }
+
+        return story
       }
     )
