+++ src/lib/index/searchIndex.ts
@@ -33,7 +33,7 @@ export type SearchIndexResponse<T> = SearchIndexError | SearchIndexResult<T>
  * @returns Promise<SearchIndexResponse>
  */
 export async function searchIndex<T>(search: object, options: SearchIndexOptions, page: number = 1, size: number = 100): Promise<SearchIndexResponse<T>> {
-  const endpoint = new URL(`${options.index}/_search`, options.endpoint)
+  const endpoint = new URL(`${options.index.replaceAll('/', '_')}/_search`, options.endpoint)
   const { from, pageSize } = pagination({
     page,
     size
